Add files via upload
A simple quizz that I made for an assignment at University. There are 10 questions each with only one single correct answer. The questions are drawn from the Basis of Technology Information subject that i studied during the first semester. The questions and answers are all in Romanian so I'm sorry for any foreigners that are interested in my work :).